+++ Auth/__init__.py
@@ -1,0 +1,2 @@
+from Auth import AuthForm
+from Auth import AuthPage
