+++ Base/__init__.py
@@ -1,0 +1,2 @@
+from Base import Page
+from Base import Component
