+++ tests/__init__.py
@@ -1,4 +1,4 @@
-from tests.example_test import DirectoryTest
-from tests.example_test import SortAndFilterTest
-from tests.example_test import CreatingDocumentsTest
-from tests.example_test import TabsAtHomePageTest
+from tests.directories_test import DirectoryTest
+from tests.sort_and_filter_test import SortAndFilterTest
+from tests.creating_document_test import CreatingDocumentsTest
+from tests.tabs_at_home_test import TabsAtHomePageTest
